services/analytics-service/requirements.txt
# Analytics Service Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# BigQuery & Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-auth==2.23.4

# Data Processing & Analytics
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
scipy==1.11.4

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# HTTP Client & Async
httpx==0.25.2
aioredis==2.0.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Data Validation & Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-mock==0.7.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Production
gunicorn==21.2.0
requests==2.31.0
